refactor(mania-replay): 删除 ManiaReplayInputParser/Event 独立文件，内联到 ManiaReplayInputData
